Refactor panel package structure and clean up dialogs
Moved panel classes and forms into a more consistent package structure under Panels subpackages for Dashboard, ManageGuests, ManageStaff, Reservation, and RoomOperation. Updated all relevant imports in MainScreen and related files. Cleaned up and removed template comments and unused code in staff dialog classes. Adjusted layout in RegisterStaffMember dialog for improved alignment.
