+++ src/com/bookify/Interfaces/ManageStaff/Dialogs/NextRegisterStaffMember.java
@@ -1,13 +1,6 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JDialog.java to edit this template
- */
+
 package com.bookify.Interfaces.ManageStaff.Dialogs;
 
-/**
- *
- * @author WW
- */
 public class NextRegisterStaffMember extends javax.swing.JDialog {
 
   
@@ -16,11 +9,6 @@ public class NextRegisterStaffMember extends javax.swing.JDialog {
         initComponents();
     }
 
-    /**
-     * This method is called from within the constructor to initialize the form.
-     * WARNING: Do NOT modify this code. The content of this method is always
-     * regenerated by the Form Editor.
-     */
     @SuppressWarnings("unchecked")
     // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
     private void initComponents() {
@@ -41,9 +29,6 @@ public class NextRegisterStaffMember extends javax.swing.JDialog {
         pack();
     }// </editor-fold>//GEN-END:initComponents
 
-    /**
-     * @param args the command line arguments
-     */
     public static void main(String args[]) {
         /* Set the Nimbus look and feel */
         //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
